Refactored tools into module #160
