+++ plaso/cli/log2timeline_tool.py
@@ -13,14 +13,8 @@ from dfvfs.resolver import context as dfvfs_context
 
 import plaso
 
-# The following import makes sure the analyzers are registered.
-from plaso import analyzers  # pylint: disable=unused-import
-
 # The following import makes sure the filters are registered.
 from plaso import filters  # pylint: disable=unused-import
-
-# The following import makes sure the parsers are registered.
-from plaso import parsers  # pylint: disable=unused-import
 
 # The following import makes sure the output modules are registered.
 from plaso import output  # pylint: disable=unused-import
